requirements.txt
requests
aiohttp
pandas
pytz
pyyaml
python-dotenv
